Fixed blur functionality in frontend.

+++ lib/backend/api_requests/api_calls.dart
@@ -80,7 +80,7 @@ class UpscaleCall {
 class BlurCall {
   static Future<ApiCallResponse> call({
     FFUploadedFile? imagePath,
-    double? radius,
+    int? radius,
   }) async {
     return ApiManager.instance.makeApiCall(
       callName: 'Blur',
